Refactored WordleInstructions to Instructions, InstructionsUseCaseInteractor.java to InstructionsInteractor.java.
