+++ src/main/java/use_case/instructions/InstructionsOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.WordleInstructions;
+package use_case.instructions;
